Updates FruityMesh to version 0.3. Now works with S130 v2.0.0 prod and SDK11. Changes are documented in CHANGELOG.md. It is important that you read this.

+++ inc/EnrollmentModule.h
@@ -107,7 +107,7 @@ class EnrollmentModule: public Module
 
 
 	public:
-		EnrollmentModule(u16 moduleId, Node* node, ConnectionManager* cm, const char* name, u16 storageSlot);
+		EnrollmentModule(u8 moduleId, Node* node, ConnectionManager* cm, const char* name, u16 storageSlot);
 
 		void ConfigurationLoadedHandler();
