+++ src_examples/TemplateModule.cpp
@@ -30,7 +30,7 @@ extern "C"{
 
 }
 
-TemplateModule::TemplateModule(u16 moduleId, Node* node, ConnectionManager* cm, const char* name, u16 storageSlot)
+TemplateModule::TemplateModule(u8 moduleId, Node* node, ConnectionManager* cm, const char* name, u16 storageSlot)
 	: Module(moduleId, node, cm, name, storageSlot)
 {
 	//Register callbacks n' stuff
